Initialize DirectX 12 and issue a clear call.
Almost Christmas (in 4 days) and I finally took  some time to work again on this project. For me, this change is another milestone, it's the farest point I achieved on all the restart I've made of this project (at
least 8 if I recall correctly).

So the easy part is now over, and here come the unknown realm. Quite exciting for what comes next !